requirements/dev.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
better-profanity==0.7.0
bidict==0.21.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
Django==3.2.8
django-multiselectfield
django-neomodel
dnspython==2.1.0
eventlet==0.32.0
fastapi==0.70.0
Flask==2.0.1
Flask-SocketIO==5.1.1
greenlet==1.1.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
markovify==0.9.3
MarkupSafe==2.0.1
monotonic==1.6
multidict==5.2.0
oauthlib==3.1.1
priority==2.0.0
pydantic==1.8.2
PySocks==1.7.1
python-dotenv==0.19.0
python-engineio==4.2.1
python-multipart==0.0.5
python-socketio==5.4.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
tweepy==3.10.0
typing-extensions==3.10.0.2
Unidecode==1.3.0
urllib3==1.26.6
uvicorn==0.15.0
Werkzeug==2.0.1
wsproto==1.0.0
yarl==1.7.0
